+++ cmd/services.go
@@ -125,6 +125,8 @@ type APIClient interface {
 	GetApplicationDemoConfig(requestContext context.Context, applicationID string) (json.RawMessage, error)
 	UpdateApplicationDemoConfig(requestContext context.Context, applicationID string, configuration json.RawMessage) (json.RawMessage, error)
 	FixApplicationDemo(requestContext context.Context, applicationID string, workspaceID string) (json.RawMessage, error)
+	GetApplicationImageRegistry(requestContext context.Context, applicationID string) (json.RawMessage, error)
+	UpdateApplicationImageRegistry(requestContext context.Context, applicationID string, registryRequest client.UpdateApplicationImageRegistryRequest) (json.RawMessage, error)
 
 	ListOrganisations() ([]client.OrganisationSummary, error)
 	SwitchOrganisation(orgID string) (*client.SwitchOrganisationResponse, error)